Updates shared models and dependencies
Updates package versions in the .csproj file to the latest versions.

Removes the local Language enum and uses the one from the SharedModels package, promoting code reuse and consistency.

Updates namespaces for localization and user requests.

These changes ensure compatibility with the latest features of the shared models library and external dependencies.