Enhances webhook setup and certificate handling
Improves the robustness of webhook setup by adding a try-catch block to handle potential exceptions during the process.
This ensures that any errors during webhook setup are caught and reported to administrators, preventing the bot from failing silently.

Also, adds a default path for the certificate pem file and adds a check to ensure the certificate directory exists before attempting to generate certificates.
The host address for the webhook is also cleaned, to remove the https:// or http:// prefix.
Adds ACMESharpCore package to handle certificate generation.