Dockerizes the application for containerization
Adds Docker support for easier deployment and scaling.

Introduces a `docker-compose.yml` file to define the application's services, networks, and volumes.

The Dockerfile is updated to use a new entrypoint script which ensures the proper permissions for the certificate folder.
An image `imkolganov/openvpn-gate-monitor-telegrambot` will be pulled instead of building from Dockerfile.

Updates assembly version.